🚚 move lib crate to lib/ folder and bin crate to bin/ folder
When working with cargo workspaces it is a good idea to split
the different crates into a lib/ and bin/ folder. This is
desired because it makes the project flatter and easier to
search/index.

The lib/ folder contains all the libraries [lib] crates that
are in the project. The bin/ folder contains all the binaries
[bin] crates that are in the project. When a crate is not
published to crates.io, the version number will be "0.0.0".

Also the quickr_core will be renamed to quickr_qr, since 'core'
is not that semantic. All the crates in lib/ are the core for
the crate 'quickr' inside the bin/ folder.